Don't stretch the logo
The CSS before kept the logo always at least 100px tall, which meant
that, at certain page widths, the natural proportions of the image
would be too wide for the window, and so the image would have to be
condensed horizontally to fit on the page while preserving that height.

While keeping a min-height might be good for readability, I’m a big
believer in not stretching type blindly along one dimension, and
stretching something as fundamental as your logo just seems weird.

So, to not stretch it, but keep it contained in the box, we just set a
simple `width: 100%`.